AUT-439 - Pull out JWKs keys into their own service
- Decouple the JWKs keys from the TokenValidationService. There will be multiple JWKs key in the future and some lambdas will just need to get the keys without needing to pull in the token validation service.
