+++ shared/src/test/java/uk/gov/di/authentication/shared/services/TokenValidationServiceTest.java
@@ -1,15 +1,10 @@
 package uk.gov.di.authentication.shared.services;
 
-import com.amazonaws.services.kms.model.GetPublicKeyRequest;
-import com.amazonaws.services.kms.model.GetPublicKeyResult;
 import com.nimbusds.jose.JOSEException;
-import com.nimbusds.jose.JWSAlgorithm;
 import com.nimbusds.jose.JWSSigner;
 import com.nimbusds.jose.crypto.ECDSASigner;
 import com.nimbusds.jose.jwk.Curve;
 import com.nimbusds.jose.jwk.ECKey;
-import com.nimbusds.jose.jwk.JWK;
-import com.nimbusds.jose.jwk.KeyUse;
 import com.nimbusds.jose.jwk.gen.ECKeyGenerator;
 import com.nimbusds.jwt.SignedJWT;
 import com.nimbusds.oauth2.sdk.id.Subject;
@@ -20,53 +15,34 @@ import org.junit.jupiter.api.Test;
 import uk.gov.di.authentication.shared.helpers.NowHelper;
 import uk.gov.di.authentication.sharedtest.helper.TokenGeneratorHelper;
 
-import java.nio.ByteBuffer;
 import java.time.temporal.ChronoUnit;
-import java.util.Base64;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 
-import static java.util.Collections.singletonList;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static uk.gov.di.authentication.shared.helpers.HashHelper.hashSha256String;
 
 class TokenValidationServiceTest {
 
-    private final ConfigurationService configurationService = mock(ConfigurationService.class);
-    private final KmsConnectionService kmsConnectionService = mock(KmsConnectionService.class);
+    private final JwksService jwksService = mock(JwksService.class);
     private final TokenValidationService tokenValidationService =
-            new TokenValidationService(configurationService, kmsConnectionService);
+            new TokenValidationService(jwksService);
     private static final Subject SUBJECT = new Subject("some-subject");
     private static final List<String> SCOPES = List.of("openid", "email", "phone");
     private static final List<String> REFRESH_SCOPES = List.of("openid", "email", "offline_access");
     private static final String CLIENT_ID = "client-id";
-    private static final String BASE_URL = "http://example.com";
+    private static final String BASE_URL = "https://example.com";
     private static final String KEY_ID = "14342354354353";
-    private static final String HASHED_KEY_ID = hashSha256String(KEY_ID);
     private JWSSigner signer;
     private ECKey ecJWK;
 
     @BeforeEach
     void setUp() throws JOSEException {
-        Optional<String> baseUrl = Optional.of(BASE_URL);
-        when(configurationService.getOidcApiBaseURL()).thenReturn(baseUrl);
         ecJWK = generateECKeyPair();
         signer = new ECDSASigner(ecJWK);
-        when(configurationService.getTokenSigningKeyAlias()).thenReturn(KEY_ID);
-        GetPublicKeyResult getPublicKeyResult = new GetPublicKeyResult();
-        getPublicKeyResult.setKeyUsage("SIGN_VERIFY");
-        getPublicKeyResult.setKeyId(KEY_ID);
-        getPublicKeyResult.setSigningAlgorithms(singletonList(JWSAlgorithm.ES256.getName()));
-        getPublicKeyResult.setPublicKey(
-                ByteBuffer.wrap(ecJWK.toPublicJWK().toECPublicKey().getEncoded()));
-        when(kmsConnectionService.getPublicKey(any(GetPublicKeyRequest.class)))
-                .thenReturn(getPublicKeyResult);
+        when(jwksService.getPublicTokenJwkWithOpaqueId()).thenReturn(ecJWK.toPublicJWK());
     }
 
     @Test
@@ -109,31 +85,6 @@ class TokenValidationServiceTest {
         assertFalse(
                 tokenValidationService.validateRefreshTokenSignatureAndExpiry(
                         new RefreshToken(signedAccessToken.serialize())));
-    }
-
-    @Test
-    void shouldRetrievePublicKeyFromKmsAndParseToJwk() {
-        byte[] publicKey =
-                Base64.getDecoder()
-                        .decode(
-                                "MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEpRm+QZsh2IkUWcqXUhBI9ulOzO8dz0Z8HIS6m77tI4eWoZgKYUcbByshDtN4gWPql7E5mN4uCLsg5+6SDXlQcA==");
-
-        when(configurationService.getTokenSigningKeyAlias()).thenReturn(KEY_ID);
-
-        var result =
-                new GetPublicKeyResult()
-                        .withKeyUsage("SIGN_VERIFY")
-                        .withKeyId(KEY_ID)
-                        .withSigningAlgorithms(singletonList(JWSAlgorithm.ES256.getName()))
-                        .withPublicKey(ByteBuffer.wrap(publicKey));
-
-        when(kmsConnectionService.getPublicKey(any(GetPublicKeyRequest.class))).thenReturn(result);
-
-        JWK publicKeyJwk = tokenValidationService.getPublicJwkWithOpaqueId();
-
-        assertEquals(publicKeyJwk.getKeyID(), HASHED_KEY_ID);
-        assertEquals(publicKeyJwk.getAlgorithm(), JWSAlgorithm.ES256);
-        assertEquals(publicKeyJwk.getKeyUse(), KeyUse.SIGNATURE);
     }
 
     @Test
